feat: finished grpc server and finished httpapi gateway server (the code is bad but at least working:))

+++ src/lab-4/GrpcServer/Mappers/PlMapper.cs
@@ -9,7 +9,7 @@ public class PlMapper
         return new ProductCreationDto
         {
             Name = productCreationModel.Name,
-            Price = productCreationModel.Price,
+            Price = (decimal)productCreationModel.Price,
         };
     }
 }
